terms-analyzer-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.2,<3.0.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6
mistralai==0.0.7 